feat: refactor locale formatting support and improve metric icon positioning in dashboard
